+++ src/Service/CreateRequest.php
@@ -2,6 +2,7 @@
 
 namespace DerSpiegel\WoodWingAssetsClient\Service;
 
+use DerSpiegel\WoodWingAssetsClient\AssetsClient;
 use DerSpiegel\WoodWingAssetsClient\Exception\AssetsException;
 use Exception;
 
@@ -13,16 +14,39 @@ use Exception;
  */
 class CreateRequest extends CreateRequestBase
 {
+    public function __construct(
+        AssetsClient $assetsClient,
+        mixed $filedata = null,
+        array $metadata = [],
+        array $metadataToReturn = ['all'],
+        bool $parseMetadataModification = false,
+        readonly bool $autoRename = false
+    )
+    {
+        parent::__construct($assetsClient, $filedata, $metadata, $metadataToReturn, $parseMetadataModification);
+    }
+
+
     public function __invoke(): AssetResponse
     {
-        $data = self::cleanMetadata($this->metadata);
+        $requestData = [
+            'autoRename' => $this->autoRename ? 'true' : 'false',
+            'parseMetadataModifications' => $this->parseMetadataModification ? 'true' : 'false',
+            'metadataToReturn' => implode(',', $this->metadataToReturn)
+        ];
+
+        $metadata = self::cleanMetadata($this->metadata);
+
+        if (count($metadata) > 0) {
+            $requestData['metadata'] = json_encode($metadata);
+        }
 
         if (is_resource($this->filedata)) {
-            $data['Filedata'] = $this->filedata;
+            $requestData['Filedata'] = $this->filedata;
         }
 
         try {
-            $response = $this->assetsClient->serviceRequest('create', $data);
+            $response = $this->assetsClient->serviceRequest('create', $requestData);
         } catch (Exception $e) {
             throw new AssetsException(sprintf('%s: Create failed: %s', __METHOD__, $e->getMessage()), $e->getCode(), $e);
         }
@@ -32,7 +56,8 @@ class CreateRequest extends CreateRequestBase
         $this->logger->info('Asset created',
             [
                 'method' => __METHOD__,
-                'metadata' => $this->metadata
+                'assetId' => $assetResponse->id->id,
+                'metadata' => $metadata
             ]
         );