+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 6.1.0 - 2024-02-26
 
-Fix SearchAssetRequest ignoring the metadataToReturn parameter, always returning all metadata.
+Fix CreateRequest::$parseMetadataModifications and CreateRequest::$metadataToReturn parameters being ignored. Add 
+support for the CreateRequest::$autoRename parameter.
+
+Fix SearchAssetRequest and UpdateRequest ignoring the $metadataToReturn parameter, always returning all metadata.
 
 ## 6.0.0 - 2024-01-10