+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 6.1.0 - 2024-02-26
+
+Fix CreateRequest::$parseMetadataModifications and CreateRequest::$metadataToReturn parameters being ignored. Add 
+support for the CreateRequest::$autoRename parameter.
+
+Fix SearchAssetRequest and UpdateRequest ignoring the $metadataToReturn parameter, always returning all metadata.
+
 ## 6.0.0 - 2024-01-10
 
 Backward incompatible API change: Introducing the AssetId class. Asset IDs must now be passed as an AssetId instance